+++ src/ai_marketplace_monitor/email_notify.py
@@ -14,25 +14,39 @@ from jinja2 import Environment, FileSystemLoader, select_autoescape
 
 from .ai import AIResponse  # type: ignore
 from .listing import Listing
-from .utils import BaseConfig, NotificationStatus, fetch_with_retry, hilight, resize_image_data
+from .notification import NotificationConfig, NotificationStatus
+from .utils import fetch_with_retry, hilight, resize_image_data
 
 
 @dataclass
-class SMTPConfig(BaseConfig):
+class EmailNotificationConfig(NotificationConfig):
+    email: List[str] | None = None
     smtp_server: str | None = None
     smtp_port: int | None = None
     smtp_username: str | None = None
     smtp_password: str | None = None
     smtp_from: str | None = None
 
-    def handle_smtp_server(self: "SMTPConfig") -> None:
+    def handle_email(self: "EmailNotificationConfig") -> None:
+        if self.email is None:
+            return
+        if isinstance(self.email, str):
+            self.email = [self.email]
+        if not isinstance(self.email, list) or not all(
+            (isinstance(x, str) and "@" in x and "." in x.split("@")[1]) for x in self.email
+        ):
+            raise ValueError(
+                f"Item {hilight(self.name)} email must be a string or list of string."
+            )
+
+    def handle_smtp_server(self: "EmailNotificationConfig") -> None:
         if self.smtp_server is None:
             return
         if not isinstance(self.smtp_server, str):
             raise ValueError("user requires a string smtp_server.")
         self.smtp_server = self.smtp_server.strip()
 
-    def handle_smtp_port(self: "SMTPConfig") -> None:
+    def handle_smtp_port(self: "EmailNotificationConfig") -> None:
         if self.smtp_port is None:
             return
 
@@ -41,7 +55,7 @@ class SMTPConfig(BaseConfig):
         if self.smtp_port < 1 or self.smtp_port > 65535:
             raise ValueError("user requires an integer smtp_port between 1 and 65535.")
 
-    def handle_smtp_username(self: "SMTPConfig") -> None:
+    def handle_smtp_username(self: "EmailNotificationConfig") -> None:
         if self.smtp_username is None:
             return
         # smtp_username should be a string
@@ -49,7 +63,7 @@ class SMTPConfig(BaseConfig):
             raise ValueError("user requires a string smtp_username.")
         self.smtp_username = self.smtp_username.strip()
 
-    def handle_smtp_password(self: "SMTPConfig") -> None:
+    def handle_smtp_password(self: "EmailNotificationConfig") -> None:
         if self.smtp_password is None:
             return
         # smtp_password should be a string
@@ -57,7 +71,7 @@ class SMTPConfig(BaseConfig):
             raise ValueError("user requires a string smtp_password.")
         self.smtp_password = self.smtp_password.strip()
 
-    def handle_smtp_from(self: "SMTPConfig") -> None:
+    def handle_smtp_from(self: "EmailNotificationConfig") -> None:
         if self.smtp_from is None:
             return
         # smtp_from should be a string
@@ -66,7 +80,7 @@ class SMTPConfig(BaseConfig):
         self.smtp_from = self.smtp_from.strip()
 
     def get_title(
-        self: "SMTPConfig",
+        self: "EmailNotificationConfig",
         listings: List[Listing],
         notification_status: List[NotificationStatus],
         force: bool = False,
@@ -97,7 +111,7 @@ class SMTPConfig(BaseConfig):
         return title
 
     def get_text_message(
-        self: "SMTPConfig",
+        self: "EmailNotificationConfig",
         listings: List[Listing],
         ratings: List[AIResponse],
         notification_status: List[NotificationStatus],
@@ -134,7 +148,7 @@ class SMTPConfig(BaseConfig):
         return message
 
     def get_html_message(
-        self: "SMTPConfig",
+        self: "EmailNotificationConfig",
         listings: List[Listing],
         ratings: List[AIResponse],
         notification_status: List[NotificationStatus],
@@ -187,15 +201,14 @@ class SMTPConfig(BaseConfig):
         return html, images
 
     def notify_through_email(
-        self: "SMTPConfig",
-        recipients: List[str] | None,
+        self: "EmailNotificationConfig",
         listings: List[Listing],
         ratings: List[AIResponse],
         notification_status: List[NotificationStatus],
         force: bool = False,
         logger: Logger | None = None,
     ) -> bool:
-        if not recipients:
+        if not self.email:
             if logger:
                 logger.debug("No recipients specified. No email sent.")
             return False
@@ -211,25 +224,22 @@ class SMTPConfig(BaseConfig):
         html_message, images = self.get_html_message(
             listings, ratings, notification_status, force, logger=logger
         )
-        return self.send_email_message(
-            recipients, title, message, html_message, images, logger=logger
-        )
+        return self.send_email_message(title, message, html_message, images, logger=logger)
 
     def send_email_message(
-        self: "SMTPConfig",
-        recipients: List[str] | None,
+        self: "EmailNotificationConfig",
         title: str,
         message: str,
         html: str,
         images: List[Tuple[bytes, str, str]],
         logger: Logger | None = None,
     ) -> bool:
-        if not recipients:
+        if not self.email:
             if logger:
                 logger.debug("No recipients specified. No email sent.")
             return False
 
-        sender = self.smtp_from or self.smtp_username or recipients[0]
+        sender = self.smtp_from or self.smtp_username or self.email[0]
 
         if self.smtp_server:
             smtp_server = self.smtp_server
@@ -241,7 +251,7 @@ class SMTPConfig(BaseConfig):
         msg["Subject"] = title
         # can use the humanized version of self.name as well
         msg["From"] = formataddr(("AI Marketplace Monitor", sender))
-        msg["To"] = ", ".join(recipients)
+        msg["To"] = ", ".join(self.email)
 
         # Create alternative part
         alt_part = MIMEMultipart("alternative")